refactor: 결제 금액 Long -> BigDecimal로 변경

+++ src/main/java/roomescape/application/dto/request/PaymentRequest.java
@@ -1,4 +1,6 @@
 package roomescape.application.dto.request;
 
-public record PaymentRequest(String paymentKey, String orderId, Long amount) {
+import java.math.BigDecimal;
+
+public record PaymentRequest(String paymentKey, String orderId, BigDecimal amount) {
 }
